Rewrite IterMut.
The last iteration contains a mutable reference, so we had to performs
lots of alias checking just to avoid undefined behavior, without real benefits.

This version uses pointers to remove those checks.

Some uncessary unsafe marks are removed, because as long as the pointer
points to valid objects, there will be no undefined behaviors even if
the iterators' size is wrong.